api-gateway/requirements.txt
uvicorn[standard]
ariadne
fastapi
aiokafka
asyncpg
python-dotenv
